Definitive Version for Conference Experiment

+++ Assets/Scripts/BeliefMethods/IMapHierarchicalBelief.cs
@@ -4,5 +4,5 @@ using System.Collections;
 public interface IMapHierarchicalBelief : IMapBelief {
 
     //Path<MapSquare> ExpandPath(Path<MapSquare> path);
-    void OpenOldPortals(float timeLimit);
+    void OpenOldPortals(long timeLimit);
 }
